Use ClickHouse plugin only for logs
The ClickHouse plugin can now only be used as interface for the
kobsio/fluent-bit-clickhouse plugin. To run raw SQL queries against a
ClickHouse instance within kobs, the new SQL plugin from #149 can be
used.
