+++ plugins/clickhouse/src/components/panel/Panel.tsx
@@ -3,7 +3,6 @@ import React, { memo } from 'react';
 import { IPluginPanelProps, PluginOptionsMissing } from '@kobsio/plugin-core';
 import { IPanelOptions } from '../../utils/interfaces';
 import Logs from './Logs';
-import SQL from './SQL';
 
 interface IPanelProps extends IPluginPanelProps {
   options?: IPanelOptions;
@@ -16,38 +15,19 @@ export const Panel: React.FunctionComponent<IPanelProps> = ({
   times,
   options,
 }: IPanelProps) => {
-  if (
-    !options ||
-    !times ||
-    (options.type === 'logs' &&
-      (!options.queries || !Array.isArray(options.queries) || options.queries.length === 0)) ||
-    (options.type === 'sql' && (!options.queries || !Array.isArray(options.queries) || options.queries.length === 0))
-  ) {
+  if (!options || !times || !options.type) {
     return (
       <PluginOptionsMissing
         title={title}
         message="Options for ClickHouse panel are missing or invalid"
         details="The panel doesn't contain the required options to render get the ClickHouse data or the provided options are invalid."
-        documentation="https://kobs.io/plugins/elasticsearch"
+        documentation="https://kobs.io/plugins/clickhouse"
       />
     );
   }
 
   if (options.type === 'logs' && options.queries) {
-    return (
-      <Logs
-        name={name}
-        title={title}
-        description={description}
-        queries={options.queries}
-        showChart={options.showChart || false}
-        times={times}
-      />
-    );
-  }
-
-  if (options.type === 'sql' && options.queries) {
-    return <SQL name={name} title={title} description={description} queries={options.queries} />;
+    return <Logs name={name} title={title} description={description} queries={options.queries} times={times} />;
   }
 
   return null;
